feat: 添加多个测试文件以增强模型推理功能，并在 Gemma4 视觉模型中添加调试输出

+++ src/ops/kernel/f32/scalar.rs
@@ -61,12 +61,13 @@ pub fn forward_q8_rows_scalar(
     );
     for out_idx in start..end {
         let row_off = out_idx * n_in;
-        let mut sum = 0.0f32;
+        let mut sum = 0.0f64;
         for block in 0..n_in.div_ceil(32) {
             let input_scale = input_scales[block];
             let block_end = ((block + 1) * 32).min(n_in);
             for in_idx in (block * 32)..block_end {
-                sum += weight[row_off + in_idx] * (input_q8[in_idx] as i8 as f32) * input_scale;
+                sum += (weight[row_off + in_idx] * (input_q8[in_idx] as i8 as f32) * input_scale)
+                    as f64;
             }
         }
         let output_index = if output.len() >= n_out {
@@ -74,7 +75,7 @@ pub fn forward_q8_rows_scalar(
         } else {
             out_idx - start
         };
-        output[output_index] = sum;
+        output[output_index] = sum as f32;
     }
 }
